workspace - checa se diretório existe

+++ provision/0000-workspace.sh
@@ -9,5 +9,11 @@ fi
 homeDir=$( grep "^$provUser" /etc/passwd | cut -d ":" -f6 )
 
 sudo usermod -aG vboxsf $provUser
-sudo -iu $provUser mkdir -p /media/sf_storage/workspace
-sudo -iu $provUser ln -s /media/sf_storage/workspace $homeDir/
+
+if [ ! -e /media/sf_storage/workspace ]; then
+    sudo -iu $provUser mkdir -p /media/sf_storage/workspace
+fi
+
+if [ ! -e $homeDir/workspace ]; then
+    sudo -iu $provUser ln -s /media/sf_storage/workspace $homeDir/
+fi
